move constants to openai
